break main into different scripts & user interface to ba able to retry invalid move & prep agent functions
