+++ Dockerfiles/Dockerfile_openai
@@ -1,0 +1,54 @@
+# Copyright (c) Jupyter Development Team.
+# Distributed under the terms of the Modified BSD License.
+ARG BASE_CONTAINER=jupyter/scipy-notebook
+FROM $BASE_CONTAINER
+
+LABEL maintainer="Jupyter Project <jupyter@googlegroups.com>"
+
+# Install Tensorflow
+# RUN conda install --quiet --yes \
+#     'tensorflow=1.12*' \
+#     'keras=2.2*' && \
+#     conda clean -tipsy && \
+#     fix-permissions $CONDA_DIR && \
+#     fix-permissions /home/$NB_USER
+
+RUN pip install tqdm==4.31.1
+RUN pip install xgboost==0.82
+RUN pip install python-Levenshtein==0.12.0
+RUN pip install flatten_json==0.1.6
+RUN pip install tqdm==4.36.1
+RUN pip install shap==0.28.5
+RUN pip install logmatic-python==0.1.7
+RUN pip install backoff
+
+RUN pip install graphene==2.1.3
+RUN pip install streamlit==0.51.0
+RUN pip install bokeh==1.4.0
+RUN pip install pyspark==2.4.4
+RUN pip install PyQt5==5.14.1
+
+USER root
+
+RUN \
+  apt-get update && \
+  apt-get install -y openjdk-8-jdk && \
+  rm -rf /var/lib/apt/lists/*
+
+
+ADD https://github.com/krallin/tini/releases/download/v0.18.0/tini /tini
+RUN chmod +x /tini
+
+
+# RUN pip install jupyter
+
+ENV JUPYTER_TOKEN=123
+
+WORKDIR /bagh-bandi
+
+ENV PYTHONPATH /bangh-bandi
+
+EXPOSE 19999
+
+ENTRYPOINT ["/tini", "--"]
+CMD ["jupyter-notebook", "--allow-root", "--no-browser", "--port=19999"]
