+++ src/app/__main__.py
@@ -1,0 +1,11 @@
+from src.app.user_interface import UserInterface
+
+
+def main():
+    ui = UserInterface()
+    ui.play_user_vs_user()
+
+
+# main
+if __name__ == '__main__':
+    main()
